requirements.txt
fastapi
google-auth
httpx
nltk
pydantic
pyjwt
pymysql
python-dotenv
redis
requests
sqlalchemy
uvicorn